# Correspondence_backend/src/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Correspondence System API Documentation",
      "description": "Fake Online REST API for Testing and Prototyping",
      "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api/",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/delete/paging/v2": {
            "get": {
                "summary": "Get all deleted threads",
                "description": "get all deleted threads",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "description": "page parameter represents the current page location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    { 
                        "in": "query",
                        "name": "limit",
                        "description": "page parameter represents the current page location",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Got all trashed threads successfully",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/delete/deleteBulkThreads": {
            "delete": {
                "summary": "Delete a list of threads",
                "description": "Given a list of thread IDs, they should be markd as deleted",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "thread_IDs",
                        "description": "An array of thread IDs to be marked as deleted",
                        "required": true,
                        "schema": {
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked all provided thread IDs as trashed",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/delete/deleteThreadById": {
            "delete": {
                "summary": "Delete a threads",
                "description": "Given a thread ID, it should be markd as deleted",
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "thread_id",
                        "description": "The ID of the thread to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked that thread as trashed",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/delete/reFill": {
            "get": {
                "summary": "Refill all deleted threads in real time asynchronously",
                "description": "Refill all deleted threads from the next page",
                "parameters": [
                    {
                        "in": "query",
                        "name": "fromComp",
                        "description": "A String denoting the name of the component from which the request is sent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "count",
                        "description": "An Integer number denoting the number of threads to fetch",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "A number denoting the size of the page",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "A number denoting the next page to be fetched from",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All of the threads to replace the deleted one",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/delete/count": {
            "get": {
                "summary": "Get the count of all trashed threads",
                "description": "Get the count of all trashed threads",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "count": {
                                    "type": "integer"
                                }
                            }
                        }
                    }, 
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "correspondence": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "corr_no": {
                    "type": "string"
                },
                "corr_type": {
                    "type": "string"
                },
                "entity_no": {
                    "type": "string"
                },
                "from_user": {
                    "type": "array"
                },
                "cc_entity": {
                    "type": "array"
                },
                "to_entity": {
                    "type": "string"
                },
                "to_department": {
                    "type": "string"
                },
                "sent_date": {
                    "type": "string"
                },
                "corr_subject": {
                    "type": "string"
                },
                "corr_body": {
                    "type": "string"
                },
                "message_status": {
                    "type": "string"
                },
                "docs_IDs": {
                    "type": "array"
                },
                "due_date": {
                    "type": "string"
                },
                "delete": {
                    "type": "boolean"
                },
                "draft": {
                    "type": "boolean"
                },
                "thread_id": {
                    "type": "string"
                },
                "await_reply": {
                    "type": "boolean"
                },
                "classification": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "starred": {
                    "type": "boolean"
                },
                "isThreadStarred": {
                    "type": "boolean"
                },
                "isThreadDrafted": {
                    "type": "boolean"
                }
            }
        }
    }
  }